IOS-2385: Add logging to crashlytics
  - Add collecting logs to crashlytics
