+++ Tangem/App/Services/Common/Analytics/Analytics.swift
@@ -1,0 +1,238 @@
+//
+//  Analytics.swift
+//  Tangem
+//
+//  Created by Alexander Osokin on 31.03.2020.
+//  Copyright © 2020 Smart Cash AG. All rights reserved.
+//
+
+import Foundation
+import FirebaseAnalytics
+import FirebaseCrashlytics
+import AppsFlyerLib
+import BlockchainSdk
+import Amplitude
+import TangemSdk
+
+enum Analytics {
+    static func log(_ event: Event, params: [ParameterKey: String] = [:]) {
+        let compatibles = event.analyticsSystems()
+        for compatible in compatibles {
+            switch compatible {
+            case .appsflyer, .firebase:
+                log(event: event, with: params)
+            case .amplitude: // TODO: Enable in future release
+                let convertedParams = params.reduce(into: [:]) { $0[$1.key.rawValue] = $1.value }
+                logAmplitude(event: event, params: convertedParams)
+            }
+        }
+        logCrashlytics(event, with: (params as [ParameterKey: Any]).firebaseParams)
+    }
+
+    static func log(event: Event, with params: [ParameterKey: Any]? = nil) {
+        let key = event.rawValue
+        let values = params?.firebaseParams
+        FirebaseAnalytics.Analytics.logEvent(key, parameters: values)
+        AppsFlyerLib.shared().logEvent(key, withValues: values)
+    }
+
+    static func logScan(card: Card, config: UserWalletConfig) {
+        log(event: .cardIsScanned, with: collectCardData(card))
+
+        if DemoUtil().isDemoCard(cardId: card.cardId) {
+            log(event: .demoActivated, with: [.cardId: card.cardId])
+        }
+    }
+
+    static func logTx(blockchainName: String?, isPushed: Bool = false) {
+        log(event: isPushed ? .transactionIsPushed : .transactionIsSent,
+            with: [ParameterKey.blockchain: blockchainName ?? ""])
+    }
+
+    static func logCardSdkError(_ error: TangemSdkError, for action: Action, parameters: [ParameterKey: Any] = [:]) {
+        if case .userCancelled = error { return }
+
+        var params = parameters
+        params[.action] = action.rawValue
+        params[.errorKey] = String(describing: error)
+
+        let nsError = NSError(domain: "Tangem SDK Error #\(error.code)", code: error.code, userInfo: params.firebaseParams)
+        Crashlytics.crashlytics().record(error: nsError)
+    }
+
+    static func logCardSdkError(_ error: TangemSdkError, for action: Action, card: Card, parameters: [ParameterKey: Any] = [:]) {
+        logCardSdkError(error, for: action, parameters: collectCardData(card, additionalParams: parameters))
+    }
+
+    static func log(error: Error) {
+        if case .userCancelled = error.toTangemSdkError() {
+            return
+        }
+
+        if let detailedDescription = (error as? DetailedError)?.detailedDescription {
+            var params = [ParameterKey: Any]()
+            params[.errorDescription] = detailedDescription
+            let nsError = NSError(domain: "DetailedError",
+                                  code: 1,
+                                  userInfo: params.firebaseParams)
+            Crashlytics.crashlytics().record(error: nsError)
+        } else {
+            Crashlytics.crashlytics().record(error: error)
+        }
+    }
+    
+    static func logCrashlytics(_ event: Event, with params: [String: Any] = [:]) {
+        if AppEnvironment.isXcodePreview {
+            return
+        }
+
+        let message = "\(event.rawValue).\(params.isEmpty ? "" : " \(params.description)")"
+        Crashlytics.crashlytics().log(message)
+    }
+
+    static func logWcEvent(_ event: WalletConnectEvent) {
+        var params = [ParameterKey: Any]()
+        let firEvent: Event
+        switch event {
+        case let .error(error, action):
+            if let action = action {
+                params[.walletConnectAction] = action.rawValue
+            }
+            params[.errorDescription] = error.localizedDescription
+            let nsError = NSError(domain: "WalletConnect Error for: \(action?.rawValue ?? "WC Service error")", code: 0, userInfo: params.firebaseParams)
+            Crashlytics.crashlytics().record(error: nsError)
+            return
+        case .action(let action):
+            params[.walletConnectAction] = action.rawValue
+            firEvent = .wcSuccessResponse
+        case .invalidRequest(let json):
+            params[.walletConnectRequest] = json
+            firEvent = .wcInvalidRequest
+        case .session(let state, let url):
+            switch state {
+            case .connect:
+                firEvent = .wcNewSession
+            case .disconnect:
+                firEvent = .wcSessionDisconnected
+            }
+            params[.walletConnectDappUrl] = url.absoluteString
+        }
+
+        log(event: firEvent, with: params)
+    }
+
+    static func logShopifyOrder(_ order: Order) {
+        var appsFlyerDiscountParams: [String: Any] = [:]
+        var firebaseDiscountParams: [String: Any] = [:]
+
+        if let discountCode = order.discount?.code {
+            appsFlyerDiscountParams[AFEventParamCouponCode] = discountCode
+            firebaseDiscountParams[AnalyticsParameterCoupon] = discountCode
+        }
+
+        let sku = order.lineItems.first?.sku ?? "unknown"
+
+        AppsFlyerLib.shared().logEvent(AFEventPurchase, withValues: appsFlyerDiscountParams.merging([
+            AFEventParamContentId: sku,
+            AFEventParamRevenue: order.total,
+            AFEventParamCurrency: order.currencyCode,
+        ], uniquingKeysWith: { $1 }))
+
+        FirebaseAnalytics.Analytics.logEvent(AnalyticsEventPurchase, parameters: firebaseDiscountParams.merging([
+            AnalyticsParameterItems: [
+                [AnalyticsParameterItemID: sku],
+            ],
+            AnalyticsParameterValue: order.total,
+            AnalyticsParameterCurrency: order.currencyCode,
+        ], uniquingKeysWith: { $1 }))
+
+        logAmplitude(event: .purchased, params: ["SKU": sku, "Count": "\(order.lineItems.count)", "Amount": "\(order.total)\(order.currencyCode)"])
+    }
+
+    static func logAmplitude(event: Event, params: [String: String] = [:]) {
+        if !AppEnvironment.isDebug {
+            Amplitude.instance().logEvent(event.rawValue, withEventProperties: params)
+        }
+    }
+
+    private static func collectCardData(_ card: Card, additionalParams: [ParameterKey: Any] = [:]) -> [ParameterKey: Any] {
+        var params = additionalParams
+        params[.batchId] = card.batchId
+        params[.firmware] = card.firmwareVersion.stringValue
+        params[.currency] = card.walletCurves.reduce("", { $0 + $1.rawValue })
+        return params
+    }
+}
+
+extension Analytics {
+    enum Action: String {
+        case scan = "tap_scan_task"
+        case sendTx = "send_transaction"
+        case pushTx = "push_transaction"
+        case walletConnectSign = "wallet_connect_personal_sign"
+        case walletConnectTxSend = "wallet_connect_tx_sign"
+        case readPinSettings = "read_pin_settings"
+        case changeSecOptions = "change_sec_options"
+        case createWallet = "create_wallet"
+        case purgeWallet = "purge_wallet"
+        case deriveKeys = "derive_keys"
+        case preparePrimary = "prepare_primary"
+        case readPrimary = "read_primary"
+        case addbackup = "add_backup"
+        case proceedBackup = "proceed_backup"
+    }
+
+    enum ParameterKey: String {
+        case blockchain = "blockchain"
+        case batchId = "batch_id"
+        case firmware = "firmware"
+        case action = "action"
+        case errorDescription = "error_description"
+        case errorCode = "error_code"
+        case newSecOption = "new_security_option"
+        case errorKey = "Tangem SDK error key"
+        case walletConnectAction = "wallet_connect_action"
+        case walletConnectRequest = "wallet_connect_request"
+        case walletConnectDappUrl = "wallet_connect_dapp_url"
+        case currencyCode = "currency_code"
+        case source = "source"
+        case cardId = "cardId"
+        case tokenName = "token_name"
+        case type
+        case currency = "Currency Type"
+        case success
+        case token = "Token"
+        case mode = "Mode"
+        case state = "State"
+        case basicCurrency = "Currency"
+        case batch = "Batch"
+    }
+
+    enum ParameterValue: String {
+        case welcome
+        case walletOnboarding = "wallet_onboarding"
+    }
+
+    enum AnalyticSystem {
+        case firebase
+        case amplitude
+        case appsflyer
+    }
+
+    enum WalletConnectEvent {
+        enum SessionEvent {
+            case disconnect
+            case connect
+        }
+
+        case error(Error, WalletConnectAction?), session(SessionEvent, URL), action(WalletConnectAction), invalidRequest(json: String?)
+    }
+}
+
+fileprivate extension Dictionary where Key == Analytics.ParameterKey, Value == Any {
+    var firebaseParams: [String: Any] {
+        var convertedParams = [String: Any]()
+        forEach { convertedParams[$0.key.rawValue] = $0.value }
+        return convertedParams
+    }
+}
